WIP: example_bundle rename EventSubscriber to EventListener
